# src/main/resources/openapi.yaml
openapi: 3.0.0
info:
  title: Filtered Mattermost API Specs
  version: 1.0.0
  description: |
    A filtered subset of the Mattermost API specification based on user criteria.
    This includes APIs for:
    - User management (finding, messaging)
    - Group, Channel, and Team management (finding, creating, managing members)
    - Messaging (posting, replying)
    For complete and recursively resolved schema definitions, please refer to the original openapi.yaml file.
servers:
  - url: "{your-mattermost-url}"
    variables:
      your-mattermost-url:
        default: http://localhost:8065
paths:
  /api/v4/users/{user_id}:
    get:
      tags:
        - users
      summary: Get a user
      description: |
        Get a user a object. Sensitive information will be sanitized out.
        ##### Permissions
        Requires an active session but no other permissions.
      operationId: GetUser
      parameters:
        - name: user_id
          in: path
          description: User GUID. This can also be "me" which will point to the current user.
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User retrieval successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "404":
          $ref: '#/components/responses/NotFound'
  /api/v4/users/username/{username}:
    get:
      tags:
        - users
      summary: Get a user by username
      description: |
        Get a user object by providing a username. Sensitive information will be sanitized out.
        ##### Permissions
        Requires an active session but no other permissions.
      operationId: GetUserByUsername
      parameters:
        - name: username
          in: path
          description: Username
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User retrieval successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "404":
          $ref: '#/components/responses/NotFound'
  /api/v4/users/email/{email}:
    get:
      tags:
        - users
      summary: Get a user by email
      description: |
        Get a user object by providing a user email. Sensitive information will be sanitized out.
        ##### Permissions
        Requires an active session and for the current session to be able to view another user's email based on the server's privacy settings.
      operationId: GetUserByEmail
      parameters:
        - name: email
          in: path
          description: User Email
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User retrieval successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
  /api/v4/users/ids:
    post:
      tags:
        - users
      summary: Get users by ids
      description: |
        Get a list of users based on a provided list of user ids.
        ##### Permissions
        Requires an active session but no other permissions.
      operationId: GetUsersByIds
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                type: string
        description: List of user ids
        required: true
      responses:
        "200":
          description: User list retrieval successful
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
  /api/v4/users/usernames:
    post:
      tags:
        - users
      summary: Get users by usernames
      description: |
        Get a list of users based on a provided list of usernames.
        ##### Permissions
        Requires an active session but no other permissions.
      operationId: GetUsersByUsernames
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                type: string
        description: List of usernames
        required: true
      responses:
        "200":
          description: User list retrieval successful
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
  /api/v4/users/search:
    post:
      tags:
        - users
      summary: Search users
      description: |
        Get a list of users based on search criteria provided in the request body. Searches are typically done against username, full name, nickname and email unless otherwise configured by the server.
        ##### Permissions
        Requires an active session and `read_channel` and/or `view_team` permissions for any channels or teams specified in the request body.
      operationId: SearchUsers
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - term
              properties:
                term:
                  description: The term to match against username, full name, nickname and email
                  type: string
                team_id:
                  description: If provided, only search users on this team
                  type: string
                not_in_team_id:
                  description: If provided, only search users not on this team
                  type: string
                in_channel_id:
                  description: If provided, only search users in this channel
                  type: string
                not_in_channel_id:
                  description: If provided, only search users not in this channel. Must specifiy `team_id` when using this option
                  type: string
                in_group_id:
                  description: If provided, only search users in this group. Must have `manage_system` permission.
                  type: string
                group_constrained:
                  description: When used with `not_in_channel_id` or `not_in_team_id`, returns only the users that are allowed to join the channel or team based on its group constrains.
                  type: boolean
                allow_inactive:
                  description: When `true`, include deactivated users in the results
                  type: boolean
                without_team:
                  type: boolean
                  description: Set this to `true` if you would like to search for users that are not on a team. This option takes precendence over `team_id`, `in_channel_id`, and `not_in_channel_id`.
                limit:
                  description: |
                    The maximum number of users to return in the results
                    __Available as of server version 5.6. Defaults to `100` if not provided or on an earlier server version.__
                  type: integer
                  default: 100
        description: Search criteria
        required: true
      responses:
        "200":
          description: User list retrieval successful
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
  /api/v4/channels/direct:
    post:
      tags:
        - channels
      summary: Create a direct message channel
      description: |
        Create a new direct message channel between two users.
        ##### Permissions
        Must be one of the two users and have `create_direct_channel` permission. Having the `manage_system` permission voids the previous requirements.
      operationId: CreateDirectChannel
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                type: string
              minItems: 2
              maxItems: 2
        description: The two user ids to be in the direct message
        required: true
      responses:
        "201":
          description: Direct channel creation successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Channel'
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
  /api/v4/posts:
    post:
      tags:
        - posts
      summary: Create a post
      description: |
        Create a new post in a channel. To create the post as a comment on another post, provide `root_id`.
        ##### Permissions
        Must have `create_post` permission for the channel the post is being created in.
      operationId: CreatePost
      parameters:
        - name: set_online
          in: query
          description: Whether to set the user status as online or not.
          schema:
            type: boolean
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - channel_id
                - message
              properties:
                channel_id:
                  type: string
                  description: The channel ID to post in
                message:
                  type: string
                  description: The message contents, can be formatted with Markdown
                root_id:
                  type: string
                  description: The post ID to comment on
                file_ids:
                  type: array
                  description: A list of file IDs to associate with the post. Note that posts are limited to 5 files maximum. Please use additional posts for more files.
                  items:
                    type: string
                props:
                  description: A general JSON property bag to attach to the post
                  type: object
                metadata:
                  description: A JSON object to add post metadata, e.g the post's priority
                  type: object
        description: Post object to create
        required: true
      responses:
        "201":
          description: Post creation successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
  /api/v4/groups:
    get:
      tags:
        - groups
      summary: Get groups
      description: |
        Retrieve a list of all groups not associated to a particular channel or team.
        If you use `not_associated_to_team`, you must be a team admin for that particular team (permission to manage that team).
        If you use `not_associated_to_channel`, you must be a channel admin for that particular channel (permission to manage that channel).
        __Minimum server version__: 5.11
      operationId: GetGroups
      parameters:
        - name: page
          in: query
          description: The page to select.
          schema:
            type: integer
            default: 0
        - name: per_page
          in: query
          description: The number of groups per page.
          schema:
            type: integer
            default: 60
        # Other parameters omitted for brevity, see full spec for details
      responses:
        "200":
          description: Group list retrieval successful
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Group'
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
  /api/v4/channels/{channel_id}:
    get:
      tags:
        - channels
      summary: Get a channel
      description: |
        Get channel from the provided channel id string.
        ##### Permissions
        `read_channel` permission for the channel.
      operationId: GetChannel
      parameters:
        - name: channel_id
          in: path
          description: Channel GUID
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Channel retrieval successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Channel'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
  /api/v4/teams/{team_id}/channels/name/{channel_name}:
    get:
      tags:
        - channels
      summary: Get a channel by name
      description: |
        Gets channel from the provided team id and channel name strings.
        ##### Permissions
        `read_channel` permission for the channel.
      operationId: GetChannelByName
      parameters:
        - name: team_id
          in: path
          description: Team GUID
          required: true
          schema:
            type: string
        - name: channel_name
          in: path
          description: Channel Name
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Channel retrieval successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Channel'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
  /api/v4/channels/search:
    post:
      tags:
        - channels
      summary: Search all private and open type channels across all teams
      description: |
        Returns all private and open type channels where 'term' matches on the name, display name, or purpose of the channel.
      operationId: SearchAllChannels
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - term
              properties:
                term:
                  type: string
                  description: The string to search in the channel name, display name, and purpose.
                # Other properties omitted for brevity
        description: The search terms and logic to use in the search.
        required: true
      responses:
        "200":
          description: Paginated channel response.
          content:
            application/json:
              schema:
                type: object # Simplified for brevity, actual response is more complex
                properties:
                  channels:
                    type: array
                    items:
                      $ref: '#/components/schemas/Channel'
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
  /api/v4/teams/{team_id}:
    get:
      tags:
        - teams
      summary: Get a team
      description: |
        Get a team on the system.
        ##### Permissions
        Must be authenticated and have the `view_team` permission.
      operationId: GetTeam
      parameters:
        - name: team_id
          in: path
          description: Team GUID
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Team retrieval successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Team'
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
  /api/v4/teams/name/{name}:
    get:
      tags:
        - teams
      summary: Get a team by name
      description: |
        Get a team based on provided name string
        ##### Permissions
        Must be authenticated, team type is open and have the `view_team` permission.
      operationId: GetTeamByName
      parameters:
        - name: name
          in: path
          description: Team Name
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Team retrieval successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Team'
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
  /api/v4/teams/search:
    post:
      tags:
        - teams
      summary: Search teams
      description: |
        Search teams based on search term and options provided in the request body.
        ##### Permissions
        Logged in user only shows open teams
        Logged in user with "manage_system" permission shows all teams
      operationId: SearchTeams
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                term:
                  description: The search term to match against the name or display name of teams
                  type: string
                # Other properties omitted for brevity
        description: Search criteria
        required: true
      responses:
        "200":
          description: Paginated teams response.
          content:
            application/json:
              schema:
                type: object # Simplified for brevity
                properties:
                  teams:
                    type: array
                    items:
                      $ref: '#/components/schemas/Team'
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
  /api/v4/users/{user_id}/teams:
    get:
      tags:
        - teams
      summary: Get a user's teams
      description: |
        Get a list of teams that a user is on.
        ##### Permissions
        Must be authenticated as the user or have the `manage_system` permission.
      operationId: GetTeamsForUser
      parameters:
        - name: user_id
          in: path
          description: User GUID
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Team list retrieval successful
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Team'
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
  /api/v4/channels/{channel_id}/members:
    get:
      tags:
        - channels
      summary: Get channel members
      description: |
        Get a page of members for a channel.
        ##### Permissions
        `read_channel` permission for the channel.
      operationId: GetChannelMembers
      parameters:
        - name: channel_id
          in: path
          description: Channel GUID
          required: true
          schema:
            type: string
        - name: page
          in: query
          description: The page to select.
          schema:
            type: integer
            default: 0
        - name: per_page
          in: query
          description: The number of members per page.
          schema:
            type: integer
            default: 60
      responses:
        "200":
          description: Channel members retrieval successful
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ChannelMember'
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
    post:
      tags:
        - channels
      summary: Add user(s) to channel
      description: Add a user(s) to a channel by creating a channel member object(s).
      operationId: AddChannelMember
      parameters:
        - name: channel_id
          in: path
          description: The channel ID
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                user_id:
                  type: string
                  description: The ID of user to add into the channel, for backwards compatibility.
                user_ids:
                  type: array
                  items:
                    type: string
                  minItems: 1
                  maxItems: 1000
                  description: The IDs of users to add into the channel, required if 'user_id' doess not exist.
                post_root_id:
                  type: string
                  description: The ID of root post where link to add channel member originates
        required: true
      responses:
        "201":
          description: Channel member creation successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChannelMember'
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
  /api/v4/teams/{team_id}/members:
    get:
      tags:
        - teams
      summary: Get team members
      description: |
        Get a page team members list based on query string parameters - team id, page and per page.
        ##### Permissions
        Must be authenticated and have the `view_team` permission.
      operationId: GetTeamMembers
      parameters:
        - name: team_id
          in: path
          description: Team GUID
          required: true
          schema:
            type: string
        - name: page
          in: query
          description: The page to select.
          schema:
            type: integer
            default: 0
        - name: per_page
          in: query
          description: The number of users per page.
          schema:
            type: integer
            default: 60
      responses:
        "200":
          description: Team members retrieval successful
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TeamMember'
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
    post:
      tags:
        - teams
      summary: Add user to team
      description: |
        Add user to the team by user_id.
        ##### Permissions
        Must be authenticated and team be open to add self. For adding another user, authenticated user must have the `add_user_to_team` permission.
      operationId: AddTeamMember
      parameters:
        - name: team_id
          in: path
          description: Team GUID
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                team_id:
                  type: string
                user_id:
                  type: string
        required: true
      responses:
        "201":
          description: Team member creation successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamMember'
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
  /api/v4/channels/group:
    post:
      tags:
        - channels
      summary: Create a group message channel
      description: |
        Create a new group message channel to group of users. If the logged in user's id is not included in the list, it will be appended to the end.
        ##### Permissions
        Must have `create_group_channel` permission.
      operationId: CreateGroupChannel
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                type: string
              minItems: 3 # MaxItems is 8 in full spec
        description: User ids to be in the group message channel.
        required: true
      responses:
        "201":
          description: Group channel creation successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Channel'
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
...
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
        create_at:
          description: The time in milliseconds a user was created
          type: integer
          format: int64
        update_at:
          description: The time in milliseconds a user was last updated
          type: integer
          format: int64
        delete_at:
          description: The time in milliseconds a user was deleted
          type: integer
          format: int64
        username:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        nickname:
          type: string
        email:
          type: string
        email_verified:
          type: boolean
        auth_service:
          type: string
        roles:
          type: string
        locale:
          type: string
        notify_props:
          $ref: '#/components/schemas/UserNotifyProps'
        props:
          type: object
        last_password_update:
          type: integer
          format: int64
        last_picture_update:
          type: integer
          format: int64
        failed_attempts:
          type: integer
        mfa_active:
          type: boolean
        timezone:
          $ref: '#/components/schemas/Timezone'
        terms_of_service_id:
          description: ID of accepted terms of service, if any. This field is not present if empty.
          type: string
        terms_of_service_create_at:
          description: The time in milliseconds the user accepted the terms of service
          type: integer
          format: int64
...
  responses:
    BadRequest:
      description: Invalid or missing parameters in URL or request body
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AppError'
...